+++ car-racing-tournament-api/Controllers/RaceController.cs
@@ -14,6 +14,7 @@ namespace car_racing_tournament_api.Controllers
         private IDriver _driverService;
         private ITeam _teamService;
         private Interfaces.IResult _resultService;
+        private ISeason _seasonService;
         private IConfiguration _configuration;
 
         public RaceController(
@@ -22,6 +23,7 @@ namespace car_racing_tournament_api.Controllers
             IDriver driverService, 
             ITeam teamService, 
             Interfaces.IResult resultService,
+            ISeason seasonService,
             IConfiguration configuration)
         {
             _raceService = raceService;
@@ -29,6 +31,7 @@ namespace car_racing_tournament_api.Controllers
             _driverService = driverService;
             _teamService = teamService;
             _resultService = resultService;
+            _seasonService = seasonService;
             _configuration = configuration;
         }
 
@@ -52,9 +55,13 @@ namespace car_racing_tournament_api.Controllers
             if (!await _permissionService.IsAdminModerator(new Guid(User.Identity!.Name!), resultGet.Race!.SeasonId))
                 return Forbid();
 
-            /*if (resultGet.Race.Season.IsArchived) {
+            var resultGetSeason = await _seasonService.GetSeasonById(resultGet.Race.SeasonId);
+            if (!resultGetSeason.IsSuccess)
+                return NotFound(resultGetSeason.ErrorMessage);
+
+            if (resultGetSeason.Season!.IsArchived) {
                 return BadRequest(_configuration["ErrorMessages:SeasonArchived"]);
-            }*/
+            }
             
             var resultUpdate = await _raceService.UpdateRace(resultGet.Race, raceDto);
             if (!resultUpdate.IsSuccess)
@@ -73,9 +80,13 @@ namespace car_racing_tournament_api.Controllers
             if (!await _permissionService.IsAdminModerator(new Guid(User.Identity!.Name!), resultGet.Race!.SeasonId))
                 return Forbid();
 
-            /*if (resultGet.Race.Season.IsArchived) {
+            var resultGetSeason = await _seasonService.GetSeasonById(resultGet.Race.SeasonId);
+            if (!resultGetSeason.IsSuccess)
+                return NotFound(resultGetSeason.ErrorMessage);
+
+            if (resultGetSeason.Season!.IsArchived) {
                 return BadRequest(_configuration["ErrorMessages:SeasonArchived"]);
-            }*/
+            }
 
             var result = await _raceService.DeleteRace(resultGet.Race);
             if (!result.IsSuccess)
@@ -108,9 +119,13 @@ namespace car_racing_tournament_api.Controllers
             if (!await _permissionService.IsAdminModerator(new Guid(User.Identity!.Name!), resultGetRace.Race!.SeasonId))
                 return Forbid();
 
-            // if (resultGetRace.Race.Season.IsArchived) {
-            //     return BadRequest(_configuration["ErrorMessages:SeasonArchived"]);
-            // }
+            var resultGetSeason = await _seasonService.GetSeasonById(resultGetRace.Race.SeasonId);
+            if (!resultGetSeason.IsSuccess)
+                return NotFound(resultGetSeason.ErrorMessage);
+
+            if (resultGetSeason.Season!.IsArchived) {
+                return BadRequest(_configuration["ErrorMessages:SeasonArchived"]);
+            }
 
             var resultGetDriver = await _driverService.GetDriverById(resultDto.DriverId);
             if (!resultGetDriver.IsSuccess)
